fix: infinite loop when keeping v-model format different than string
By default flatpick always outputs a string no matter what the input format
is provided. But in custom implementations we might want to keep same i/o formats
by implementing our on @input callback instead of the default v-model.

The problem:
The watcher logic is performing strict equality which doesn't detect same
non-string input values, that causes an infinite loop because for the watcher
the value is never equal. Also the loop isn't triggered in the beginning
because the watcher is only called after the first value mutation.

Related issue:
https://github.com/ankurk91/vue-flatpickr-component/issues/190